pyproject.toml
[tool.poetry]
name = "life-agent"
version = "0.1.0"
description = "Backend for LifeMap using FastAPI and Google Gemini"
authors = ["LifeMap Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
google-generativeai = "^0.3.2"  # Verifying latest version compatibility is good, but sticking to reasonable defaults
pydantic-settings = "^2.1.0"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.25"
pgvector = "^0.2.4" # Helpful for vector ops with sqlalchemy if needed, though raw SQL or specialized libs might be used.

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = "^0.1.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I"] # Enable pycodestyle errors, pyflakes, and isort
